functions enabled
functions interface
view save_ubication.dart
view ubication_recolect.dart
